@@ -20,7 +20,6 @@ import (
 	"archive/tar"
 	"bytes"
 	"compress/gzip"
-	"context"
 	"encoding/json"
 	"fmt"
 	"math/rand"
@@ -45,7 +44,7 @@ func TestContentLimit(t *testing.T) {
 		t.Skip("skipping test because existing deployment is used")
 	}
 
-	ctx := context.Background()
+	ctx := t.Context()
 
 	var b util.Binary
 	tmpDir := util.CreateTestDir(t, "content-limit-test")
@@ -70,6 +69,7 @@ func TestContentLimit(t *testing.T) {
 
 	stdout, stderr, err := b.Run(
 		"import",
+		"--no-encryption",
 		"-d", dumpPath,
 		"--pmm-url", pmm.PMMURL())
 	if err != nil {
@@ -84,6 +84,7 @@ func TestContentLimit(t *testing.T) {
 
 	stdout, stderr, err = b.Run(
 		"import",
+		"--no-encryption",
 		"-d", dumpPath,
 		"--pmm-url", pmm.PMMURL(),
 		"--vm-content-limit", "10024")
@@ -135,7 +136,7 @@ func generateFakeDump(filepath string) error {
 		return fmt.Errorf("failed to write dump meta content: %w", err)
 	}
 
-	for i := 0; i < 10; i++ {
+	for i := range 10 {
 		content, err := generateFakeChunk(100000)
 		if err != nil {
 			return fmt.Errorf("failed to generate fake chunk: %w", err)
@@ -164,7 +165,7 @@ func generateFakeDump(filepath string) error {
 func generateFakeChunk(size int) ([]byte, error) {
 	r := rand.New(rand.NewSource(time.Now().Unix())) //nolint:gosec
 	var data []byte
-	for i := 0; i < size; i++ {
+	for i := range size {
 		metricsData, err := json.Marshal(victoriametrics.Metric{
 			Metric: map[string]string{
 				"__name__": "test",
