@@ -20,17 +20,18 @@ import (
 	"archive/tar"
 	"bytes"
 	"compress/gzip"
-	"context"
 	"crypto/sha256"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
 	"math"
+	"math/big"
 	"os"
 	"path"
 	"path/filepath"
 	"reflect"
+	"strconv"
 	"strings"
 	"testing"
 	"time"
@@ -42,8 +43,15 @@ import (
 	"pmm-dump/pkg/victoriametrics"
 )
 
+const (
+	chunkTimeRange       = time.Second * 30
+	timeRange            = time.Second * 120
+	precisionForRounding = 20
+	bitsForRounding      = 64
+)
+
 func TestValidate(t *testing.T) {
-	ctx := context.Background()
+	ctx := t.Context()
 
 	c := deployment.NewController(t)
 	pmm := c.NewPMM("validate", ".env.test")
@@ -53,15 +61,14 @@ func TestValidate(t *testing.T) {
 	tmpDir := util.CreateTestDir(t, "validate-test")
 	xDumpPath := filepath.Join(tmpDir, "dump.tar.gz")
 	yDumpPath := filepath.Join(tmpDir, "dump2.tar.gz")
-	chunkTimeRange := time.Second * 30
 
 	if err := pmm.Deploy(ctx); err != nil {
 		t.Fatal(err)
 	}
 
 	start := time.Now().UTC()
-	pmm.Log("Sleeping for 120 seconds")
-	time.Sleep(time.Second * 120)
+	pmm.Logf("Sleeping for %d seconds", int(timeRange.Seconds()))
+	time.Sleep(timeRange)
 	end := time.Now().UTC()
 
 	pmm.Log("Exporting data to", xDumpPath, start, end)
@@ -74,7 +81,9 @@ func TestValidate(t *testing.T) {
 		"--click-house-url", pmm.ClickhouseURL(),
 		"--start-ts", start.Format(time.RFC3339),
 		"--end-ts", end.Format(time.RFC3339),
-		"--chunk-time-range", chunkTimeRange.String())
+		"--chunk-time-range", chunkTimeRange.String(),
+		"--no-encryption",
+		"-v")
 	if err != nil {
 		t.Fatal("failed to export", err, stdout, stderr)
 	}
@@ -92,7 +101,9 @@ func TestValidate(t *testing.T) {
 		"-d", xDumpPath,
 		"--pmm-url", newPMM.PMMURL(),
 		"--dump-qan",
-		"--click-house-url", newPMM.ClickhouseURL())
+		"--click-house-url", newPMM.ClickhouseURL(),
+		"--no-encryption",
+		"-v")
 	if err != nil {
 		t.Fatal("failed to import", err, stdout, stderr)
 	}
@@ -108,8 +119,11 @@ func TestValidate(t *testing.T) {
 		"--pmm-url", newPMM.PMMURL(),
 		"--dump-qan",
 		"--click-house-url", newPMM.ClickhouseURL(),
-		"--start-ts", start.Format(time.RFC3339), "--end-ts", end.Format(time.RFC3339),
-		"--chunk-time-range", chunkTimeRange.String())
+		"--start-ts", start.Format(time.RFC3339),
+		"--end-ts", end.Format(time.RFC3339),
+		"--chunk-time-range", chunkTimeRange.String(),
+		"--no-encryption",
+		"-v")
 	if err != nil {
 		t.Fatal("failed to import", err, stdout, stderr)
 	}
@@ -118,6 +132,16 @@ func TestValidate(t *testing.T) {
 	if err != nil {
 		t.Fatal("failed to validate chunks", err)
 	}
+
+	// TODO: This should be a removed and replaced with a better solution.
+	// Investigate why, after PMM 3.3.0,
+	// Postgres began creating duplicate metrics with a value of 0.
+	dublicateLoss, notDuplicateLoss, err := countDuplicateLoss(t, pmm, xDumpPath, yDumpPath)
+	if err != nil {
+		t.Fatal("failed to count duplicate loss", err)
+	}
+	loss -= (dublicateLoss - notDuplicateLoss)
+
 	if loss > 0.001 {
 		t.Fatalf("too much data loss %f%%", loss*100)
 	}
@@ -289,7 +313,6 @@ func vmCompareChunkData(pmm *deployment.PMM, xChunk, yChunk []vmMetric) (int, er
 		currentLoss := xMetric.CompareTimestampValues(pmm, yMetric)
 		if currentLoss > 0 {
 			loss += currentLoss
-			continue
 		}
 
 		delete(xHashMap, k)
@@ -370,7 +393,10 @@ func isGzip(data []byte) bool {
 	reader := bytes.NewReader(data)
 	r, err := gzip.NewReader(reader)
 	if r != nil {
-		r.Close()
+		err = r.Close()
+		if err != nil {
+			panic(err)
+		}
 	}
 	return err == nil
 }
@@ -402,7 +428,7 @@ type vmMetric victoriametrics.Metric
 
 func (vm vmMetric) MetricString() string {
 	m := vm.Metric
-	return fmt.Sprintf(`{"__name__": "%s", "job": "%s", "instance": "%s", "agent_id": "%s", "agent_type": "%s"}`, m["__name__"], m["job"], m["instance"], m["agent_id"], m["agent_type"])
+	return fmt.Sprintf(`{"__name__": "%s", "job": "%s", "instance": "%s", "agent_id": "%s", "agent_type": "%s"} (%s)`, m["__name__"], m["job"], m["instance"], m["agent_id"], m["agent_type"], fmt.Sprint(m))
 }
 
 func (vm vmMetric) Hash() string {
@@ -434,10 +460,18 @@ func (vm vmMetric) CompareTimestampValues(pmm *deployment.PMM, with vmMetric) in
 	for timestamp, xValue := range xMap {
 		yValue, ok := yMap[timestamp]
 		if !ok {
+			switch vm.Metric["__name__"] {
+			case "pg_stat_activity_count", "pg_stat_activity_max_tx_duration", "pg_stat_activity_max_state_duration":
+				if vm.Metric["datname"] == "grafana" || vm.Metric["datname"] == "pmm-managed" {
+					// Temporary fix for duplicates
+					// See [countDuplicateLoss] for details about postgres and duplicates.
+					continue
+				}
+			}
 			pmm.Log(fmt.Sprintf("Value and timestamp not found for metric %s in second dump: wanted %v for %d", vm.MetricString(), xValue, timestamp))
 			continue
 		}
-		if xValue != yValue {
+		if !roundFloatAndCheck(xValue, yValue) {
 			pmm.Log(fmt.Sprintf("Values for timestamp %d in metric %s are not the same: %v and %v", timestamp, vm.MetricString(), xValue, yValue))
 			continue
 		}
@@ -454,4 +488,172 @@ func (vm vmMetric) CompareTimestampValues(pmm *deployment.PMM, with vmMetric) in
 	}
 
 	return int(math.Abs(float64(len(xMap) - len(yMap))))
+}
+
+// Go usually can't handle floats with 16+ digits.
+// For example, when comparing them, the dump has the same numbers, but Go's float64 can interpret them as two different numbers:
+// 1) 0.00999923005927 and 0.00999923005928.
+// 2) 1.7540397181656854e+09 and 1.7540397181656857e+09.
+// 3) 9.223372036854776e+18 and 9.223372036854775e+18.
+// So, we convert them to Big.Float, round them, and then compare them.
+
+func roundFloatAndCheck(f1, f2 float64) bool {
+	// stop rounding and just compare if f1 has 1 digit only
+	s := strconv.FormatFloat(f1, 'e', precisionForRounding, bitsForRounding)
+	dotIndex := strings.Index(s, "e")
+	if dotIndex == -1 {
+		return f1 == f2
+	}
+
+	// round and compare
+	n1Rounded := new(big.Float).SetPrec(precisionForRounding).SetMode(big.ToNearestAway).SetFloat64(f1)
+	n2Rounded := new(big.Float).SetPrec(precisionForRounding).SetMode(big.ToNearestAway).SetFloat64(f2)
+
+	return n1Rounded.Cmp(n2Rounded) == 0
+}
+
+// After version 3.3.0, PostgreSQL metrics (pg_stat_activity_count, pg_stat_activity_max_tx_duration,
+// and pg_stat_activity_max_state_duration specifically from the Grafana and PMM-managed databases)
+// started showing duplicate metrics with the same timestamp.
+//
+// The only difference is that the duplicate metrics have a value of 0, and an opposite usename.
+// for example, we have two metrics from the first dump. Metrics with "##" will be missing after import.
+//
+// 1:
+//
+// Metric name: pg_stat_activity_count , value:0 timestamp:1, datname: pmm-managed, state:active, usename:
+// ##Metric name: pg_stat_activity_count , value:0 timestamp:2, datname: pmm-managed, state:active, usename:
+// ##Metric name: pg_stat_activity_count , value:0 timestamp:3, datname: pmm-managed, state:active, usename:
+// Metric name: pg_stat_activity_count , value:0 timestamp:4, datname: pmm-managed, state:active, usename:
+//
+// 2:
+//
+// ##Metric name: pg_stat_activity_count , value:0 timestamp:1, datname: pmm-managed, state:active, usename: pmm-managed
+// Metric name: pg_stat_activity_count , value:1 timestamp:2, datname: pmm-managed, state:active, usename: pmm-managed
+// Metric name: pg_stat_activity_count , value:1 timestamp:3, datname: pmm-managed, state:active, usename: pmm-managed
+// ##Metric name: pg_stat_activity_count , value:0 timestamp:4, datname: pmm-managed, state:active, usename: pmm-managed
+//
+// After importing the dump into another PMM, the metrics will be deduplicated:
+//
+// Metric name: pg_stat_activity_count , value:0 timestamp:1, datname: pmm-managed, state:active, usename:
+// Metric name: pg_stat_activity_count , value:1 timestamp:2, datname: pmm-managed, state:active, usename: pmm-managed
+// Metric name: pg_stat_activity_count , value:1 timestamp:3, datname: pmm-managed, state:active, usename: pmm-managed
+// Metric name: pg_stat_activity_count , value:0 timestamp:4, datname: pmm-managed, state:active, usename:
+//
+// Although these results are two different metrics, they were merged in comment for demonstration.
+// So, to prevent the test from failing, we need to go over two dumps again and calculate loss value of duplicates.
+// And then subtract them from actual value.
+func countDuplicateLoss(t *testing.T, pmm *deployment.PMM, xDump, yDump string) (float64, float64, error) {
+	t.Helper()
+	xChunkMap, err := readChunks(xDump)
+	if err != nil {
+		return 0, 0, fmt.Errorf("failed to read dump %s: %w", xDump, err)
+	}
+	yChunkMap, err := readChunks(yDump)
+	if err != nil {
+		return 0, 0, fmt.Errorf("failed to read dump %s: %w", yDump, err)
+	}
+
+	totalDuplicate, totalNotDuplicate, totalValues := 0, 0, 0
+
+	duplicateMapX := make(map[string]float64)
+	// only needed to compare metrics later to find the duplicates in the X dump.
+	duplicateMapY := make(map[string]float64)
+
+	checkIfPGMetric := func(vm vmMetric) bool {
+		switch vm.Metric["__name__"] {
+		case "pg_stat_activity_count", "pg_stat_activity_max_tx_duration", "pg_stat_activity_max_state_duration":
+			if vm.Metric["datname"] == "grafana" || vm.Metric["datname"] == "pmm-managed" {
+				return true
+			}
+		}
+		return false
+	}
+
+	saveTimestampToMap := func(timestamp int64, value float64, vm vmMetric, destination map[string]float64) {
+		destination[fmt.Sprintf("%s: Timestamp:%d, Datname:%s, State:%s", vm.Metric["__name__"], timestamp, vm.Metric["datname"], vm.Metric["state"])] = value
+	}
+
+	rangeOverChunk := func(xChunk, yChunk []vmMetric) {
+		xHashMap := make(map[string]vmMetric)
+		for _, v := range xChunk {
+			xHashMap[v.MetricHash()] = v
+		}
+
+		yHashMap := make(map[string]vmMetric)
+		for _, v := range yChunk {
+			yHashMap[v.MetricHash()] = v
+		}
+
+		for i, vmX := range xHashMap {
+			if !checkIfPGMetric(vmX) {
+				continue
+			}
+
+			xMap := make(map[int64]float64)
+			for i, v := range vmX.Timestamps {
+				xMap[v] = vmX.Values[i]
+			}
+
+			vmY, ok := yHashMap[i]
+			if !ok {
+				for timestamp, valueX := range xMap {
+					saveTimestampToMap(timestamp, valueX, vmX, duplicateMapX)
+				}
+			}
+
+			yMap := make(map[int64]float64)
+			for i, v := range vmY.Timestamps {
+				yMap[v] = vmY.Values[i]
+			}
+
+			for timestamp, valueX := range xMap {
+				valueY, ok := yMap[timestamp]
+				if !ok {
+					saveTimestampToMap(timestamp, valueX, vmX, duplicateMapX)
+					continue
+				}
+				saveTimestampToMap(timestamp, valueY, vmY, duplicateMapY)
+			}
+		}
+	}
+
+	for xFilename, xChunkData := range xChunkMap {
+		yChunkData := yChunkMap[xFilename]
+		dir, _ := path.Split(xFilename)
+		st := dump.ParseSourceType(dir[:len(dir)-1])
+		if st == dump.VictoriaMetrics {
+			xChunk, err := vmParseChunk(xChunkData)
+			if err != nil {
+				return 0, 0, fmt.Errorf("failed to parse chunk %s: %w", xFilename, err)
+			}
+			yChunk, err := vmParseChunk(yChunkData)
+			if err != nil {
+				return 0, 0, fmt.Errorf("failed to parse chunk %s: %w", xFilename, err)
+			}
+
+			xValues := vmValuesCount(xChunk)
+			yValues := vmValuesCount(yChunk)
+			if xValues > yValues {
+				totalValues += xValues
+			} else {
+				totalValues += yValues
+			}
+
+			rangeOverChunk(xChunk, yChunk)
+		}
+	}
+
+	for timestamp, value := range duplicateMapX {
+		_, ok := duplicateMapY[timestamp]
+		if ok {
+			totalDuplicate++
+			continue
+		}
+		pmm.Logf("Value and timestamp not found for metric %s in second dump, value:%f", timestamp, value)
+		totalNotDuplicate++
+	}
+
+	pmm.Logf("Found %d dublicates for pg_stat_activity_count, pg_stat_activity_max_tx_duration, pg_stat_activity_max_state_duration in first dump and removed them", totalDuplicate)
+	return float64(totalDuplicate) / float64(totalValues), float64(totalNotDuplicate) / float64(totalValues), nil
 }
