@@ -18,7 +18,7 @@ import "io"
 
 type Source interface {
 	Type() SourceType
-	ReadChunk(meta ChunkMeta) (*Chunk, error)
+	ReadChunks(meta ChunkMeta) ([]*Chunk, error)
 	WriteChunk(filename string, r io.Reader) error
 	FinalizeWrites() error
 }
