@@ -15,9 +15,7 @@
 package transferer
 
 import (
-	"archive/tar"
 	"bytes"
-	"compress/gzip"
 	"context"
 	"errors"
 	"fmt"
@@ -28,17 +26,17 @@ import (
 	"golang.org/x/sync/errgroup"
 
 	"pmm-dump/pkg/dump"
+	"pmm-dump/pkg/encryption"
 )
 
-func (t Transferer) Import(ctx context.Context, runtimeMeta dump.Meta) error {
+func (t Transferer) Import(ctx context.Context, runtimeMeta dump.Meta, e encryption.Options) error {
 	log.Info().Msg("Importing metrics...")
-	gzr, err := gzip.NewReader(t.file)
+	r, err := dump.NewReader(t.file, &e)
 	if err != nil {
-		return fmt.Errorf("failed to open as gzip: %w", err)
+		return fmt.Errorf("failed to create readers: %w", err)
 	}
-	defer gzr.Close() //nolint:errcheck
-
-	tr := tar.NewReader(gzr)
+	defer r.Close() //nolint:errcheck
+	tr := r.GetTarReader()
 
 	var metafileExists bool
 
