@@ -36,12 +36,12 @@ func (s fakeSource) Type() dump.SourceType {
 	return s.sourceType
 }
 
-func (s fakeSource) ReadChunk(m dump.ChunkMeta) (*dump.Chunk, error) {
-	return &dump.Chunk{
+func (s fakeSource) ReadChunks(m dump.ChunkMeta) ([]*dump.Chunk, error) {
+	return []*dump.Chunk{{
 		ChunkMeta: m,
 		Content:   []byte("content"),
 		Filename:  m.String() + ".bin",
-	}, nil
+	}}, nil
 }
 
 func (s fakeSource) WriteChunk(_ string, r io.Reader) error {
